requirements.txt
streamlit>=1.30.0
opencv-python-headless
numpy>=1.26.0
scikit-image>=0.22.0
scipy>=1.11.0
Pillow>=10.0.0
pillow-heif